requirements.txt
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
dill==0.2.6
enum34==1.1.6
future==0.16.0
futures==3.1.1
gapic-google-cloud-pubsub-v1==0.15.4
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-cloud-pubsub==0.25.0
google-gax==0.15.13
googleapis-common-protos==1.5.2
grpc-google-iam-v1==0.11.1
grpcio==1.4.0
httplib2==0.10.3
idna==2.5
oauth2client==3.0.0
ply==3.8
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
python-dateutil==2.6.0
requests==2.18.1
rsa==3.4.2
six==1.10.0
urllib3==1.21.1
